requirements.txt
python-dotenv>=1.0.0,<2.0.0
groq>=0.3.0,<0.4.0